Fix the module name is not extracted when the path to the module contains spaces
When `MainClassModuleNameExtractor` is run in external JVM a file with
arguments for the JVM is created. This file contains the paths to the
modules but if those paths contain spaces they are not properly escaped
and as a result `MainClassModuleNameExtractor` receives corrupted paths
as arguments.

Quote all paths in the file so spaces are handled properly.