Return null for module name when extraction fails
As detailed in https://issues.apache.org/jira/browse/MJAVADOC-609,
module name extraction can easily fail, due to a legacy jar having a
file naming convention that doesn't match what ModuleFinder expects,
when determining an automatic module name from the file name.

This failure to determine the automatic module name should not cause the
jar to be missing from the resulting class path.

This patch handles the FindException that ModuleFinder throws when
looking for module descriptors, and returns null, indicating no name.
This should allow the jar to still be added as a class path element,
even though it can't be added as a module path element. Named modules
won't be able to use the jar from the module path, but automatically
named modules should still be able to use the jar from the UNNAMED
module on the class path.

Also:

* Update LocationManagerIT to test the change
* Make the LocationManager perform strict check for file existence
* Fix broken LocationManagerIT tests (references to non-existent
  test/resource/dir.* directories) revealed by strict file existence
  checks in LocationManager