Fail on duplicate modules on the module path
When resolving the module path, no module id is allowed to exist
multiple times.

The location manager used to drop duplicates silently which led to
hard to understand error messages when a plugin using plexus-java
dropped a dependency without any notice.

This change enforces uniqueness of module names on the module path
and will add an exception to the PathResolveResult in that case.